added search_bar tests
