+++ ecommerce/src/pages/base_page.py
