+++ ecommerce/src/pages/home_page.py
@@ -8,5 +8,3 @@ class HomePage(BasePage):
 
     def open_home_page(self):
         self.driver.get(self.url)
-
-
