+++ ecommerce/src/pages/search_page.py
@@ -11,11 +11,24 @@ from ecommerce.src.pages.home_page import HomePage
 
 class SearchPage(HomePage):
 
+
     url = 'https://demo.nopcommerce.com/search'
   
+
     def __init__(self, driver):
         self.driver = driver 
         self._products = []
+    
+
+    @property
+    def products(self):
+        return self._products
+    
+
+    @products.setter
+    def products(self, value):
+        self._products = value
+    
     
     @property
     def page_title(self):
@@ -84,14 +97,7 @@ class SearchPage(HomePage):
 
         return products
 
-    @property
-    def products(self):
-        return self._products
-    
-    @products.setter
-    def products(self, value):
-        self._products = value
-    
+
     @property
     def pager(self, timeout=2):
         PAGER_LOCATOR = (By.XPATH, "//div[@class='pager']")
@@ -118,12 +124,11 @@ class SearchPage(HomePage):
 
         return success_notification
     
+
     @property
     def bar_notification_close(self):
         BAR_NOTIFICATION_CLOSE_LOCATOR = (By.XPATH, "//div[contains(@class, 'bar-notification success')]//span[@class='close']")
         return BaseElement(self.driver, locator=BAR_NOTIFICATION_CLOSE_LOCATOR)
-
-
 
     def wait_for_products_update(self):
         LOADING_SPINNER_LOCATOR = (By.XPATH, '//div[@class="ajax-products-busy"]')
@@ -132,8 +137,46 @@ class SearchPage(HomePage):
             self.driver, 2
             ).until(EC.invisibility_of_element_located(LOADING_SPINNER_LOCATOR))
     
+
     def open_search_page(self):
         self.open_page()
 
+
     def get_page_title(self):
         return self.page_title.text
+    
+
+    def get_search_page_title(self):
+        return self.page_title.text
+    
+
+    def is_advanced_search_selected(self):
+        advanced_search_checkbox = self.advanced_search_checkbox
+        is_selected = advanced_search_checkbox.element.is_selected()
+        return is_selected
+
+
+    def is_category_dropdown_visible(self):
+        return self.category_dropdown.is_displayed
+    
+
+    def is_search_subcategories_checkbox_visible(self):
+        return self.search_subcategories_checkbox.is_displayed
+
+
+    def is_manufacturer_dropdown_visible(self):
+        return self.manufacturer_dropdown.is_displayed
+    
+
+    def is_search_in_description_dropdown_visible(self):
+        return self.search_in_description_checkbox.is_displayed
+    
+
+    def check_advanced_search(self):
+        self.advanced_search_checkbox.click()
+    
+
+    def select_display_number(self, number):
+        if number in self.display_dropdown.get_values():
+            self.display_dropdown.select_value(number)
+            self.wait_for_products_update()
